CHROMIUM: esdfs: Fix deadlock when updating the file size on 32-bit SMP systems
On 32-bit SMP systems esdfs_setattr() acquires the inode spin lock
before updating the file size. Instead releasing the lock afterwards
with spin_unlock() it calls spin_lock() again on the same lock, which
leads to a deadlock, since the lock is already held. Release the lock
instead of trying to re-acquire it.

BUG=chromium:1000748
TEST=no more reboots due to esdfs deadlock when ARC++ starts up

Change-Id: I73b8cad2b0e47c4d19dab1ddaaa602a3cdc67f9e
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1825814
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
[5.4-arcvm: picked from 5.4 for initial merge]